requirements.txt
# Web framework
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.27
sse-starlette==3.4.1

# Database
sqlalchemy==2.0.49
aiosqlite==0.22.1

# AI
anthropic==0.99.0

# Resume parsing
pdfplumber==0.11.9
python-docx==1.2.0

# HTTP client (job APIs)
httpx==0.28.1

# Configuration
python-dotenv==1.2.2
pydantic-settings==2.9.1
